dax/cxl: add existing dc extents when probing dax region
Add existing dc extents on the device before probing dax region will
cause the creation of the dax device fail as resource cannot present
when driver is binded to the device as shown in really_probe().

We delay the processing of existing dc extents to cx region driver
probe.

Question: the guard() in cxlr_notify_extent() will cause lock issue,
removed it. Not sure whether it will cause issue or not although no
issue is observed during test.